MAJOR OVERHAUL OF THE CODE
No more global variables exist.
Everything is passed through the structures defined in structures.h.
Flow solver and gradient has been verified thoroughly.
Optimization works for gradient descent and BFGS, in the presence of shock or not.

TO BE DONE:
Convert Hessian stuff in the new structure.
Convert finite difference of some structures to always have a reference to test against.
	e.g. dRdW_FD, dRdS_FD, etc.
Convert other flow solver fluxes and time stepping schemes
	Note: since only the scalar dissipation flux has been linearized, no point in adding other fluxes.

Finally, current goal would be to be able to do one-shot optimization with this code.