Working linearization of the fixed point.
However the fixed point that uses the adjoint is way too slow compared to the dwdx and conventional method.

The dwdx method is also much slower than the conventional method. This might be due to the overhead of computing the jacobian at every iteration.

Instead, we'll try to solve the flow implicitly such that less flow iterations are required.

Started implementing LUSGS. Need the BC linearization that is not based on adoubles.